+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# [next]
+# 1.9.6
 - Fix bug "getting stuck" in `MoveToPositionAlongThePath`
+- Improvements in diagonal movements in `MoveToPositionAlongThePath`
 
 # 1.9.5
 - Update params name of `simpleAttackMelee` in Enemy.